feat(slatetoreact): export from separate file
