+++ src/index.ts
@@ -9,10 +9,6 @@ export { config as slateToDomConfig } from './config/slateToDom/default'
 export { config as payloadSlateToDomConfig } from './config/slateToDom/payload'
 export { config as slateDemoSlateToDomConfig } from './config/slateToDom/slateDemo'
 
-// slateToReact
-export { Config as SlateToReactConfig } from './config/slateToReact/types'
-export { config as slateToReactConfig } from './config/slateToReact/default'
-
 // htmlToSlate
 export { Config as HtmlToSlateConfig } from './config/htmlToSlate/types'
 export { config as htmlToSlateConfig } from './config/htmlToSlate/default'
