+++ src/react.ts
@@ -1,0 +1,6 @@
+// Serializers
+export { SlateToReact } from './serializers/slateToReact'
+
+// Configuration objects
+export { Config as SlateToReactConfig } from './config/slateToReact/types'
+export { config as slateToReactConfig } from './config/slateToReact/default'
